Changed paths: M       CHANGES.txt M       cat12.m M       cat_conf_extopts.m M       cat_io_senderrormail.m M       cat_main_gcut.m M       cat_run.m M       cat_run_job.m M       cat_run_job_APP_final.m M       cat_vol_correct_slice_scaling.m M       cat_vol_morph.m M       spm_cat12.m D       templates_animals/ape_lesser_T1.nii D       templates_animals/ape_lesser_T2.nii D       templates_animals/ape_lesser_TPM.nii D       templates_animals/ape_lesser_brainmask.nii D       templates_animals/cat_defaults_ape_greater.m D       templates_animals/cat_defaults_ape_lesser.m D       templates_animals/cat_defaults_monkey_newworld.m D       templates_animals/cat_defaults_monkey_oldworld.m D       templates_animals/monkey_newworld_T1.nii D       templates_animals/monkey_newworld_T2.nii D       templates_animals/monkey_newworld_TPM.nii D       templates_animals/monkey_newworld_brainmask.nii
(1) Changed:  Updated help of spm_cat12.
(2) Changed:  Updated main head classification in final APP increase the chance 
              to remove the scull and not the brain in species with thick muscles.
(3) Changed:  Small corrections.
(4) Deleted:  Removed species specific default files and set parameter in spm_cat12.
(5) Deleted:  Removed incomplete species templates.



git-svn-id: svn://141.35.69.214/cat12@952 9f44b166-36c4-483c-8295-4af200601734

+++ cat_vol_morph.m
@@ -69,7 +69,7 @@ function vol = cat_vol_morph(vol,action,n,vx_vol)
     error('MATLAB:cat_vol_morph:vx_vol', ...
       'Wrong vx_vol size. It has to be a 1x3 matrix.\n'); 
   end
-  no=n; n=round(n); 
+  no=n; n=round(double(n)); 
   switch lower(action)
     case {'l' 'lc' 'lo' 'labclose' 'labopen'}
       % not return in this case
